requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.2
PyJWT==2.6.0
